requirements.txt
PySide6>=6.8.2.1
pdf2image>=1.17.0
opencv-python>=4.11.0
pypdf>=5.3.1
numpy>=2.2.3
Pillow>=11.1.0
pyinstaller>=6.12.0
nuitka>=2.6.8